+++ Runtime/GLE.cs
@@ -48,7 +48,23 @@ namespace ProceduralToolkit
         /// </summary>
         public static void DrawRay(Ray ray)
         {
-            Draw.WireRay(drawLine, ray);
+            Draw.Ray(drawLine, ray);
+        }
+
+        /// <summary>
+        /// Draws a segment
+        /// </summary>
+        public static void DrawSegment2(Segment2 segment)
+        {
+            Draw.Segment2(drawLine, segment);
+        }
+
+        /// <summary>
+        /// Draws a segment
+        /// </summary>
+        public static void DrawSegment3(Segment3 segment)
+        {
+            Draw.Segment3(drawLine, segment);
         }
 
         #region DrawWireQuad